requirements.txt
alembic==1.1.0
Babel==2.7.0
blinker==1.4
Bootstrap-Flask==1.0.10
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Admin==1.5.3
Flask-BabelEx==0.9.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
flask-msearch==0.2.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.2
six==1.12.0
speaklater==1.3
SQLAlchemy==1.3.7
visitor==0.1.3
Werkzeug==0.15.5
WTForms==2.2.1
